Update Error Logger
Since I had some crashes in the recent time with the error log just clogging the space on my Pico due to its massive size, I wrote a script to check the size evrey time it wants to write a new error and to limit the size of the error log to 100 errors. This should be far enough for any debugging purposes.